--- js/get-random-number.js
@@ -1,13 +1,0 @@
-// https://developer.mozilla.org/ru/docs/Web/JavaScript/Reference/Global_Objects/Math/random
-function getRandomNumber (min, max) {
-  if (min >= 0 && max > min) {
-    min = Math.ceil(min);
-    max = Math.floor(max);
-
-    return Math.floor(Math.random() * (max - min + 1)) + min;
-  }
-
-  throw new Error('Ошибка. Неправильно заданный диапазон');
-}
-
-export {getRandomNumber};
