+++ js/main.js
@@ -1,3 +1,3 @@
 import {createPhotoMocks} from './data.js';
 
-console.log(createPhotoMocks());
+createPhotoMocks();
